+++ models.go
@@ -1,0 +1,9 @@
+package main
+
+type shortPost struct {
+	ID          string
+	Title       string
+	Date        string
+	Description string
+	Link        string
+}
